requirements.txt
fastapi
werkzeug
pdf2image
pillow
pymupdf
python-dotenv
requests
uvicorn
google-generativeai
